adcionando função frear e desligar
